feat: updating

+++ classes/views.py
@@ -5,10 +5,14 @@ from .models import Class, Subject
 from administration.models import Student, Teacher
 from student.models import Timetable
 from django.contrib import messages
-from teacher.models import Attendance
+from django.http import HttpResponseForbidden
+from django.contrib.auth.decorators import login_required
 
 
+@login_required
 def create_class(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     if request.method == 'POST':
         form = ClassForm(request.POST)
         if form.is_valid():
@@ -20,7 +24,10 @@ def create_class(request):
     return render(request, 'dashboards/all_admin_pages/create_class.html', {'form': form})
 
 
+@login_required
 def edit_class(request, class_id):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     classes = get_object_or_404(Class, id=class_id)
     if request.method == 'POST':
         form = ClassForm(request.POST, instance=classes)
@@ -33,7 +40,10 @@ def edit_class(request, class_id):
     return render(request, 'dashboards/all_admin_pages/create_class.html', {'form': form})
 
 
+@login_required
 def delete_class(request, class_id):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     classes = get_object_or_404(Class, id=class_id)
     if request.method=='POST':
         classes.delete()
@@ -42,12 +52,18 @@ def delete_class(request, class_id):
     return render(request, 'dashboards/all_admin_pages/delete_class.html', {'classes': classes})
 
 
+@login_required
 def class_list(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     classes = Class.objects.all()
     return render(request, 'dashboards/all_admin_pages/myClass.html', {'classes': classes})
 
 
+@login_required
 def create_subject(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     if request.method == 'POST':
         form = SubjectForm(request.POST)
         if form.is_valid():
@@ -59,7 +75,10 @@ def create_subject(request):
     return render(request, 'dashboards/all_admin_pages/subjects.html', {'form': form})
 
 
+@login_required
 def subject_list(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     all_subject = Subject.objects.all()
     classes = Class.objects.all()
     subject = Subject.objects.filter(classes__in = classes)
@@ -71,19 +90,28 @@ def subject_list(request):
     return render(request, 'dashboards/all_admin_pages/subject_list.html', context)
 
 
+@login_required
 def remove_subject(request, sub_id):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     subject = get_object_or_404(Subject, id=sub_id)
     subject.delete()
     messages.success(request, 'Subject successfully removed')
     return redirect('subject_list')
 
 
+@login_required
 def myClass(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     classes = Class.objects.all()
     return render(request, 'dashboards/all_admin_pages/myClass.html', {'class': classes})
 
 
+@login_required
 def classDetail(request, user_id):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden('You do not have permission to access this page.')
     classes = get_object_or_404(Class, id=user_id)
     student = Student.objects.filter(student_class=classes)
     teacher = Teacher.objects.filter(classes=classes)
